Operate network functions in special file 'requests.h' and 'requests.cpp', attach dependencies for installer_part1 only
